requirements.txt
torchtext==0.9.1
pytorch-lightning==1.2.8
transformers==4.5.1
parlai==1.1.1